🧠 refactor: tag entity 반영
issue: #24

+++ src/main/java/com/fourdays/foodage/review/dto/ReviewDto.java
@@ -12,7 +12,7 @@ public record ReviewDto(
 	Integer price,
 	Integer rating,
 	String contents,
-	int tagId,
+	Long tagId,
 	int thumbnailId,
 	// List<String> images,
 	// String lastEatenFood,
